(DiamondLightSource/hyperion#1077) Move detector_motion.py into detector package as per review comment
