+++ src/dodal/beamlines/i04.py
@@ -8,7 +8,7 @@ from dodal.devices.backlight import Backlight
 from dodal.devices.beamstop import BeamStop
 from dodal.devices.DCM import DCM
 from dodal.devices.detector import DetectorParams
-from dodal.devices.detector_motion import DetectorMotion
+from dodal.devices.detector.detector_motion import DetectorMotion
 from dodal.devices.eiger import EigerDetector
 from dodal.devices.fast_grid_scan import FastGridScan
 from dodal.devices.flux import Flux
